// modules/medical/types/medical-errors.ts
export enum MedicalErrorType {
  required = 'required',
  invalid = 'invalid',
  duplicated = 'duplicated',
  startGreaterThanEnd = 'startGreaterThanEnd',
  smallDateInterval = 'smallDateInterval',
  createFailed = 'createFailed',
  updateFailed = 'updateFailed',
  confirmFailed = 'confirmFailed',
  isPast = 'isPast',
  incomplete = 'incomplete',
  companyNotFound = 'companyNotFound',
  validateIdnp = 'validateIdnp',
}